backend/pyproject.toml
[tool.poetry]
name = "prior2do-backend"
version = "0.1.0"
description = ""
authors = ["Ben Choy <97584068+bench352@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.88.0"}
neo4j = "^5.3.0"
pendulum = "^2.1.2"
bcrypt = "^4.0.1"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
